+++ addOns/authhelper/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Changed
 - Ignore non-displayed fields when selecting the user name and password.
+- Use single displayed field for user name, e.g. multi step login.
 
 ## [0.17.0] - 2025-01-09
 ### Changed